==> 01_Descriptives.ipynb <==
# Descriptive Statistics for Study 3

import pandas as pd
import numpy as np
from scipy.stats import chi2_contingency
from IPython.display import display

#Allowing each cell to display multiple outputs
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = 'all'

df=pd.read_csv(r"C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\df_alls.csv")

#Examining frequencies of combinations of different suicide-related outcomes
df[['si', 'nssi', 'att']].value_counts().reset_index(name='count')

#Examining prior SITBs
#SITB variables were initially coded as 1 Yes 2 No; recode to 0 No 1 Yes
#Also recode missing values so there aren't multiple types of missing values (previously had -9 and -3)
for col in df[['hs54a',	'hs54b',	'hs54c',	'hs54d',	'hs54f']]:
    df[col]=df[col].replace({2:0})
    df[col]=df[col].mask(df[col] <0, -9)
#SI
#Creating the SI variable using variables measured in Wave 7 as defined in the paper and in data_clean notebook
#hs54c- seriously consider suicide, hs54d- suicide plan
df['si7']=df['hs54c']
df['si7']=np.where(df['hs54d']==1, 1, df['si7'])
#Renaming attempt variable and recoding it so that it becomes a binary variable as previously defined in data_clean and paper
df['att7']=df['hs54e'].replace([2,3,4],1)
#Creating the NSSI variable 
df['nssi7']=df['hs54b']


#Generating crosstabs and running the chi2 test on the prior SITB variables
sre7=['si7', 'nssi7', 'att7']
df['att7']=df['att7'].mask(df['att7'] <0, -9)

for s in sre7:
    locals()[f'{s}_cross']=pd.crosstab(index=df[s], columns=df['sitbs'], dropna=False, margins=True)
    locals()[f'{s}_cross']
    pd.crosstab(index=df[s], columns=df['sitbs'], dropna=False, normalize='columns')*100
    stat, p, dof, expected=chi2_contingency(locals()[f'{s}_cross'])
    print(f'chi2 p-value= {p}')

#Create function that produces cross-tabs and chi2 p-value
#Note that the display function is used instead of print so that the output is a nicely formatted table
def desc(var):
    locals()[f'{var}_cross']=pd.crosstab(index=df[var], columns=df['sitbs'], dropna=False, margins=True)
    display(locals()[f'{var}_cross'])
    display(pd.crosstab(index=df[var], columns=df['sitbs'], dropna=False, normalize='columns')*100)
    stat, p, dof, expected=chi2_contingency(locals()[f'{var}_cross'])
    print(f'chi sq statistic={stat}')
    print(f'p-value for chi2 test= {p}')


sexcont=pd.crosstab(index=df['f02m1_2.0'], columns=df['sitbs'], dropna=False)
sexcont
stat, p, dof, expected=chi2_contingency(sexcont)
stat
p

adhdcont=pd.crosstab(index=df['hs17l'], columns=df['sitbs'], dropna=False)
adhdcont
stat, p, dof, expected=chi2_contingency(adhdcont)
stat
p

#Sex
desc('f02m1_2.0')

#Depression/ anxiety
desc('hs17v')
#ADHD
desc('hs17l')
#Two parents who are partners
desc('parpart_2.0')
#Parent 1 has anxiety/ depression
df['p1da']=df['hs48a32'].replace([2,3,4],2)
df['p1da']=df['p1da'].mask(df['p1da'] <0, -1)
desc('p1da')
