==> 08_CI_Analyses.ipynb <==
# Further Analyses- Producing 95% CIs and Examining Most Important Features

from IPython.core.interactiveshell import InteractiveShell
from matplotlib import pyplot 
from numpy import mean
from numpy import std
from sklearn.calibration import calibration_curve
from sklearn.metrics import confusion_matrix as confusion
from sklearn.metrics import f1_score
from sklearn.metrics import precision_recall_curve
from sklearn.metrics import precision_score 
from sklearn.metrics import recall_score as recall
from sklearn.metrics import roc_auc_score
from sklearn.metrics import roc_curve
from sklearn.model_selection import cross_val_predict
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import KFold
from sklearn.model_selection import RandomizedSearchCV
from sklearn.model_selection import StratifiedKFold
from sklearn.utils import resample
import joblib
import numpy as np
import pandas as pd
import warnings
InteractiveShell.ast_node_interactivity = "all"

Bootstrapped CIs for First Round of Evals (Full Test Dataset)

#Loading datasets
Xi_hold=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\Xi_hold.csv')
y_hold_si=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_si.csv')
y_hold_nssi=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_nssi.csv')
y_hold_att=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_att.csv')
y_hold_sitbs=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_sitbs.csv')

Xi_hold.shape

#Defining a function to prepare 100 bootstrapped test sets, and obtain estimates of performance metrics for each resampled version of test data
algos=['LR', 'RF', 'XGB']
threshs1=[0.188868, 0.175579, 0.119263]
models1= {a: joblib.load(f'{a}_si.sav') for a in algos}

def boot(model, fulltest, thresh):
    #Suppressing warnings for this section. Warning appears because the bootstrapped samples do not have feature names as they are arrays while the model has previously been fitted on pandas dataframes with feature names
    #This is ok because the features are in the same order as the original test set
    warnings.simplefilter(action='ignore', category=UserWarning)
    n_iterations=100
    aucs=list()
    f1s=list()
    sens=list()
    specs=list()
    ppvs=list()
    values=fulltest.values
    for i in range(n_iterations):
        test=resample(values, n_samples=len(fulltest), stratify=values[:, -1], random_state=i)
        probs=model.predict_proba(test[:, :-1])
        probs=probs[:, 1]
        auc=roc_auc_score(test[:, -1], probs)
        aucs.append(auc)
        pred=np.where(probs>= thresh, 1, 0)
        f1=f1_score(test[:, -1], pred)
        f1s.append(f1)
        sen=recall(test[:, -1], pred)
        sens.append(sen)
        ppv=precision_score(test[:, -1], pred)
        ppvs.append(ppv)
        tn, fp, fn, tp=confusion(test[:, -1], pred).ravel()
        spec=tn/(tn+fp)
        specs.append(spec)
        
    metrics=["aucs", "f1s", "sens", "specs", "ppvs"]
    metricsdf=pd.DataFrame(zip(aucs, f1s, sens, specs, ppvs), columns=metrics)
    return metricsdf

#Iterating through tuples to evaluate LR, RF, and XGB Models 
#Value of the tuple algo is paired to the other corresponding value in threshs, i.e. 'LR' is paired with threshold of 0.188868
def runevals(models1, fulltest, threshs):
    for a, t in zip(algos, threshs):
        clf=models1[a]
        print(f'95% CIs for performance using {a} with threshold at {t}')
        metricsdf=boot(clf, fulltest, t)
        print(metricsdf.quantile([.025, .975]).round(3))
        print('\n')

fulltest_si=Xi_hold.join(y_hold_si)
runevals(models1, fulltest_si, threshs1)

threshs2=[0.138621, 0.160344, 0.172887]
models2= {a: joblib.load(f'{a}_nssi.sav') for a in algos}

fulltest_nssi=Xi_hold.join(y_hold_nssi)
runevals(models2, fulltest_nssi, threshs2)

threshs3=[0.074037, 0.073374, 0.087420]
models3= {a: joblib.load(f'{a}_att.sav') for a in algos}

fulltest_att=Xi_hold.join(y_hold_att)
runevals(models3, fulltest_att, threshs3)

threshs4=[0.208806, 0.152130, 0.176893]
models4= {a: joblib.load(f'{a}_sitbs.sav') for a in algos}

fulltest_sitbs=Xi_hold.join(y_hold_sitbs)
runevals(models4, fulltest_sitbs, threshs4)

Bootstrapped CIs for Second Round Evals (Reduced Test Dataset)

#Loading the reduced datasets created in the evaluation notebook
si_test=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\si_test.csv')
nssi_test=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\nssi_test.csv')
att_test=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\att_test.csv')

si_test=si_test.drop(columns=['nssi', 'att', 'sitbs'])
nssi_test=nssi_test.drop(columns=['si', 'att', 'sitbs'])
att_test=att_test.drop(columns=['si', 'nssi', 'sitbs'])

#Evaluating models for predicting SI
runevals(models1, si_test, threshs1)
runevals(models4, si_test, threshs1)

#Predicting NSSI
runevals(models2, nssi_test, threshs2)
runevals(models4, nssi_test, threshs2)

#Predicting attempts
runevals(models3, att_test, threshs3)
runevals(models4, att_test, threshs3)
