==> 09_Model_Evals.ipynb <==
# Model Evaluation of Performance on Individual Outcomes
This notebook compares and evaluates the performance of previously fitted models in predicting individual outcomes. Unlike the previous model evaluations which used the entire sample, these evaluations use reduced datasets comprising of only indidivudals with the outcome of interest, and a control group with no suicide-related outcomes at all. These reduced datasets circumvent problems with calculating metrics such as specificity and precision for individual outcomes in the full data.

from IPython.core.interactiveshell import InteractiveShell
from matplotlib import pyplot 
from numpy import mean
from numpy import std
from sklearn.calibration import calibration_curve
from sklearn.metrics import confusion_matrix as confusion
from sklearn.metrics import f1_score
from sklearn.metrics import precision_recall_curve
from sklearn.metrics import precision_score as ppv
from sklearn.metrics import recall_score as recall
from sklearn.metrics import roc_auc_score
from sklearn.metrics import roc_curve
from sklearn.model_selection import cross_val_predict
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import KFold
from sklearn.model_selection import RandomizedSearchCV
from sklearn.model_selection import StratifiedKFold
import joblib
import numpy as np
import pandas as pd
InteractiveShell.ast_node_interactivity = "all"

Suicidal ideation model evaluation

#Recreating full test dataset
Xi_hold=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\Xi_hold.csv')
y_hold=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold.csv')
Xi_hold.shape
y_hold.shape
testdata=Xi_hold.join(y_hold)
testdata

#Examining frequencies of the different combinations of outcomes
testdata[['si', 'nssi', 'att']].value_counts().reset_index(name='count')
#Creating the different subsets of data for evaluating performance on individual outcomes
si_test=testdata[(testdata['si']==1) | ((testdata['att']==0) & (testdata['nssi']==0))]
nssi_test=testdata[(testdata['nssi']==1) | ((testdata['att']==0) & (testdata['si']==0))]
att_test=testdata[(testdata['att']==1) | ((testdata['nssi']==0) & (testdata['si']==0))]

#Checking to see if the datasets were created properly
sre=['si', 'att', 'nssi']
for i in sre:
    print(f'{i}_test frequencies by outcome')
    #Retrieves variable value and concatenates 'i' and '_test'. locals() function is used to construct variable names dynamically
    locals()[f'{i}_test'][['si', 'nssi', 'att']].value_counts().reset_index(name='count')

#Splitting the reduced datasets back into x and y 
#y_holds_si and so on are created just for these evaluations, as opposed to y_hold_si, which contains the outcome of si for the full sample 
#s standing for small, indicating a reduced dataset
for i in sre:
    locals()[f'Xi_holds_{i}']=locals()[f'{i}_test'].drop(columns=['si', 'att', 'nssi','sitbs'])
    locals()[f'y_holds_{i}']=locals()[f'{i}_test'][f'{i}']

for i in sre:
    locals()[f'{i}_test'].to_csv(f'{i}_test.csv', index=False)

for i in sre:
    locals()[f'{i}_test'].shape

Evaluating the models' ability to predict suicidal ideation

#Creating a function that predicts probabilities and evaluates the model
def eval(model, Xi_hold, y_hold, thresh):
    #Generating probability predictions
    proba=model.predict_proba(Xi_hold)
    proba=proba[:, 1]
    print('Probs: %.3f (%.3f)' % (mean(proba), std(proba)))
    #Evaluating probability predictions
    ypred=np.where(proba>thresh, 1, 0)
    f1= f1_score(y_hold, ypred)
    print(f'F1= {f1:.3f}')
    sens= recall(y_hold, ypred)
    print(f'Sensitivity= {sens:.3f}')
    tn, fp, fn, tp= confusion(y_hold, ypred).ravel()
    spec=tn/(tn+fp)
    print(f'Specificity= {spec:.3f}')
    auc= roc_auc_score(y_hold, proba)
    print(f'AUROC= {auc:.3f}')
    prec=ppv(y_hold, ypred)
    print(f'Precision= {prec:.3f}')

    return f1, sens, spec, auc, prec

#Creating dictionary of component and composite models
algos=['LR', 'RF', 'XGB']
threshs=[0.188868, 0.175579, 0.119263]
models1= {a: joblib.load(f'{a}_si.sav') for a in algos}
models2 = {a: joblib.load(f'{a}_sitbs.sav') for a in algos}


#Iterating through tuples to evaluate LR, RF, and XGB Models 
#Value of the tuple algo is paired to the other corresponding value in threshs, i.e. 'LR' is paired with threshold of 0.188868
def runevals(models1, Xi_hold, y_hold):
    for a, t in zip(algos, threshs):
        clf=models1[a]
        print(f'Metrics for {a} using threshold at {t}')
        eval(clf, Xi_hold, y_hold, t)
        print('\n')

runevals(models1, Xi_holds_si, y_holds_si)

runevals(models2, Xi_holds_si, y_holds_si)

Evaluating the models' ability to predict NSSI

threshs=[0.138621, 0.160344, 0.172887]
models1= {a: joblib.load(f'{a}_nssi.sav') for a in algos}

runevals(models1, Xi_holds_nssi, y_holds_nssi)

runevals(models2, Xi_holds_nssi, y_holds_nssi)

Evaluating the models' ability to predict attempts

threshs=[0.074037, 0.073374, 0.087420]
models1= {a: joblib.load(f'{a}_att.sav') for a in algos}

runevals(models1, Xi_holds_att, y_holds_att)

runevals(models2, Xi_holds_att, y_holds_att)

Predicting attempts with nssi model

y_hold_att=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_att.csv')

threshs=[0.074037, 0.073374, 0.087420]
models1= {a: joblib.load(f'{a}_nssi.sav') for a in algos}

runevals(models1, Xi_hold, y_hold_att)

Deriving performance metrics for previously fitted models in full test dataset

#Loading datasets
y_hold_si=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_si.csv')
y_hold_nssi=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_nssi.csv')
y_hold_sitbs=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_sitbs.csv')

#SI Models
algos=['LR', 'RF', 'XGB']
threshs=[0.188868, 0.175579, 0.119263]
models= {a: joblib.load(f'{a}_si.sav') for a in algos}
runevals(models, Xi_hold, y_hold_si)

#NSSI Models
threshs=[0.138621, 0.160344, 0.172887]
models= {a: joblib.load(f'{a}_nssi.sav') for a in algos}
runevals(models, Xi_hold, y_hold_nssi)

#Attempt models
threshs=[0.074037, 0.073374, 0.087420]
models= {a: joblib.load(f'{a}_att.sav') for a in algos}

runevals(models, Xi_hold, y_hold_att)

#SITB Models
threshs=[0.208806, 0.152130, 0.176893]
models = {a: joblib.load(f'{a}_sitbs.sav') for a in algos}
runevals(models, Xi_hold, y_hold_sitbs)
