==> 10_Feature Importance.ipynb <==
# Feature Importance Values 

import joblib
import pandas as pd
from IPython.core.interactiveshell import InteractiveShell

InteractiveShell.ast_node_interactivity = "all"
model_si=joblib.load('XGB_si_Uncalibrated.sav')
model_nssi=joblib.load('XGB_nssi_Uncalibrated.sav')
model_att=joblib.load('XGB_att_Uncalibrated.sav')
model_sitbs=joblib.load('XGB_sitbs_Uncalibrated.sav')
Xi_hold=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\Xi_hold.csv')
y_hold_si=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_si.csv')
y_hold_nssi=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_nssi.csv')
y_hold_att=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_att.csv')
y_hold_sitbs=pd.read_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\y_hold_sitbs.csv')
Xi_hold.shape

def fscore(model):
    fscores=model.feature_importances_
    feature_names=Xi_hold.columns
    impt=pd.Series(fscores, index=feature_names)
    impt=impt.sort_values(ascending=False)
    impt=impt.head(20)
    print(impt)

    return impt

imptfeats = fscore(model_si)

w7dd=pd.read_csv(r"C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\w7ddsel.csv")
#Defining a function which extracts rows of the data dictionary corresponding to the most important features
def dd():
    imptcols=list(imptfeats.index.values)
    dd_feat=w7dd.loc[w7dd['Without age'].isin(imptcols)]
    return dd_feat

imptfeats.to_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\SI_imptfeats.csv')
si_feat_dd=dd()
si_feat_dd.to_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\si_feat_dd.csv')


imptfeats=fscore(model_nssi)

imptfeats.to_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\NSSI_imptfeats.csv')
nssi_feat_dd=dd()
nssi_feat_dd.to_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\nssi_feat_dd.csv')


imptfeats = fscore(model_att)

imptfeats.to_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\att_imptfeats.csv')
att_feat_dd=dd()
att_feat_dd.to_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\att_feat_dd.csv')

imptfeats= fscore(model_sitbs)

imptfeats.to_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\SITBs_imptfeats.csv')
si_feat_dd=dd()
si_feat_dd.to_csv(r'C:\Users\z5291979\OneDrive - UNSW\Documents\lsac-data\processed_data\sitbs_feat_dd.csv')
